try to fix on 1-array_iterator.c

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -10,7 +10,7 @@
  * */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t a;
+	int a;
 	if(array && action)
 	{
 		for(a = 0; a < size; a++)
